+++ api-server/api-server-common/src/storage/impls/in_memory/transactional/write.rs
@@ -19,8 +19,8 @@ use crate::storage::storage_api::{
     block_aux_data::{BlockAuxData, BlockWithExtraData},
     AmountWithDecimals, ApiServerStorageError, ApiServerStorageRead, ApiServerStorageWrite,
     BlockInfo, CoinOrTokenStatistic, Delegation, FungibleTokenData, LockedUtxo, NftWithOwner,
-    Order, PoolBlockStats, PoolDataWithExtraInfo, TransactionInfo, TransactionWithBlockInfo, Utxo,
-    UtxoWithExtraInfo,
+    Order, PoolBlockStats, PoolDataWithExtraInfo, TokenTransaction, TransactionInfo,
+    TransactionWithBlockInfo, Utxo, UtxoWithExtraInfo,
 };
 use common::{
     address::Address,
@@ -64,6 +64,13 @@ impl ApiServerStorageWrite for ApiServerInMemoryStorageTransactionalRw<'_> {
         self.transaction.del_address_transactions_above_height(block_height)
     }
 
+    async fn del_token_transactions_above_height(
+        &mut self,
+        block_height: BlockHeight,
+    ) -> Result<(), ApiServerStorageError> {
+        self.transaction.del_token_transactions_above_height(block_height)
+    }
+
     async fn set_address_balance_at_height(
         &mut self,
         address: &Address<Destination>,
@@ -102,6 +109,21 @@ impl ApiServerStorageWrite for ApiServerInMemoryStorageTransactionalRw<'_> {
     ) -> Result<(), ApiServerStorageError> {
         self.transaction
             .set_address_transactions_at_height(address, transactions, block_height)
+    }
+
+    async fn set_token_transaction_at_height(
+        &mut self,
+        token_id: TokenId,
+        tx_id: Id<Transaction>,
+        block_height: BlockHeight,
+        tx_global_index: u64,
+    ) -> Result<(), ApiServerStorageError> {
+        self.transaction.set_token_transaction_at_height(
+            token_id,
+            tx_id,
+            block_height,
+            tx_global_index,
+        )
     }
 
     async fn set_mainchain_block(
@@ -176,21 +198,21 @@ impl ApiServerStorageWrite for ApiServerInMemoryStorageTransactionalRw<'_> {
         &mut self,
         outpoint: UtxoOutPoint,
         utxo: Utxo,
-        address: &str,
+        addresses: &[&str],
         block_height: BlockHeight,
     ) -> Result<(), ApiServerStorageError> {
-        self.transaction.set_utxo_at_height(outpoint, utxo, address, block_height)
+        self.transaction.set_utxo_at_height(outpoint, utxo, addresses, block_height)
     }
 
     async fn set_locked_utxo_at_height(
         &mut self,
         outpoint: UtxoOutPoint,
         utxo: LockedUtxo,
-        address: &str,
+        addresses: &[&str],
         block_height: BlockHeight,
     ) -> Result<(), ApiServerStorageError> {
         self.transaction
-            .set_locked_utxo_at_height(outpoint, utxo, address, block_height)
+            .set_locked_utxo_at_height(outpoint, utxo, addresses, block_height)
     }
 
     async fn del_utxo_above_height(
@@ -329,6 +351,15 @@ impl ApiServerStorageRead for ApiServerInMemoryStorageTransactionalRw<'_> {
         address: &str,
     ) -> Result<Vec<Id<Transaction>>, ApiServerStorageError> {
         self.transaction.get_address_transactions(address)
+    }
+
+    async fn get_token_transactions(
+        &self,
+        token_id: TokenId,
+        len: u32,
+        tx_global_index: u64,
+    ) -> Result<Vec<TokenTransaction>, ApiServerStorageError> {
+        self.transaction.get_token_transactions(token_id, len, tx_global_index)
     }
 
     async fn get_latest_blocktimestamps(